fixed some warnings on msvc

+++ src/game/q_shared.c
@@ -1011,7 +1011,7 @@ MinGW comes with its own snprintf() which is not broken.
 =============
 */
 
-int Q_vsnprintf(char *str, size_t size, const char *format, va_list ap) {
+size_t Q_vsnprintf(char *str, size_t size, const char *format, va_list ap) {
 	int retval;
 	retval = _vsnprintf(str, size, format, ap);
 	if(retval < 0 || retval == size) {
